third_party/WebKit/LayoutTests/fast/ime/html-getinputcontext-expected.txt

This tests getInputContext() which is an additional interface for HTMLElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new InputMethodContext() threw exception TypeError: Illegal constructor.
PASS function getInputContext() { [native code] } is non-null.
PASS typeof(html.getInputContext) is "function"
PASS context instanceof InputMethodContext is true
PASS context2 instanceof InputMethodContext is true
PASS context === context2 is true
PASS successfullyParsed is true

TEST COMPLETE
